in prog

+++ backend/ohq/models.py
@@ -462,7 +462,7 @@ class Occurrence(models.Model):
     original_end = models.DateTimeField(("original end"))
     created_on = models.DateTimeField(("created on"), auto_now_add=True)
     updated_on = models.DateTimeField(("updated on"), auto_now=True)
-    interval = models.IntegerField(("interval"), blank=True, validators=[MinValueValidator(5), MaxValueValidator(60)])
+    interval = models.IntegerField(("interval"), blank=True, null=True, validators=[MinValueValidator(5), MaxValueValidator(60)])
 
     class Meta:
         verbose_name = ("occurrence")
@@ -480,23 +480,43 @@ class Occurrence(models.Model):
             self.location = event.location
 
     def save(self, *args, **kwargs):
+        # Get the old instance if this is an update
+        old_instance = None
+        if self.pk:
+            old_instance = Occurrence.objects.get(pk=self.pk)
+
         super().save(*args, **kwargs)
 
-        if self.pk: # If save is called on object update, not creation
-            self.bookings.all().delete()
+        # Only update bookings if:
+        # 1. This is a new occurrence
+        # 2. The interval has changed
+        # 3. The start or end time has changed
+        should_update_bookings = (
+            not old_instance or  # New occurrence
+            old_instance.interval != self.interval or  # Interval changed
+            old_instance.start != self.start or  # Start time changed
+            old_instance.end != self.end  # End time changed
+        )
 
-        delta = self.end - self.start
-        delta_minutes = delta.total_seconds() / 60
-        booking_count = int(delta_minutes // self.interval)
-        for i in range(booking_count):
-            booking_start = self.start + timedelta(minutes=i * self.interval)
-            booking_end = booking_start + timedelta(minutes=self.interval)
-            Booking.objects.create(
-                occurrence=self,
-                user=None,
-                start=booking_start,
-                end = booking_end,
-            )
+        if should_update_bookings and self.interval:
+            # Delete existing bookings
+            if self.pk:
+                self.bookings.all().delete()
+
+            # Create new bookings
+            delta = self.end - self.start
+            delta_minutes = delta.total_seconds() / 60
+            booking_count = int(delta_minutes // self.interval)
+            
+            for i in range(booking_count):
+                booking_start = self.start + timedelta(minutes=i * self.interval)
+                booking_end = booking_start + timedelta(minutes=self.interval)
+                Booking.objects.create(
+                    occurrence=self,
+                    user=None,
+                    start=booking_start,
+                    end=booking_end,
+                )
 
     def moved(self):
         return self.original_start != self.start or self.original_end != self.end
